backend/requirements.txt
Flask==2.2.2
pymongo==4.1.1
flask-cors==3.0.10
Werkzeug==2.2.2
flask-bcrypt==1.0.1
python-dotenv==1.0.1
